app/pyproject.toml
[project]
name = "power-automate-upload"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-identity>=1.21.0",
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.12",
    "identity>=0.9.2",
    "msal>=1.32.0",
    "passlib>=1.7.4",
    "pydantic-settings>=2.8.1",
    "pymongo>=4.11.3",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "requests>=2.32.3",
    "uvicorn>=0.34.0",
]
